requirements.txt
kivy>=2.3.0
numpy>=1.24.0
Pillow>=10.0.0
tensorflow==2.13.0
opencv-python>=4.9.0.0  # image processing for lesion segmentation
rembg>=2.0.50  # background removal for leaf extraction
onnxruntime>=1.16.0  # required backend for rembg
# Hardware interfaces (typically pre-installed on Raspberry Pi OS):
# picamera2
# RPi.GPIO
# Optional smaller runtime (prefer over full tensorflow for deployment); install one of:
# tflite-runtime==2.13.0
